Revert "[wip][fix] modify the type of pc (usize -> u32)"
This reverts commit f000d69159efca0c4a183364febf8efe5b76c5b2.

+++ src/cpu.rs
@@ -18,7 +18,7 @@ pub enum PrivilegedLevel {
 }
 
 pub struct CPU {
-    pub pc: u32,
+    pub pc: usize,
     pub regs: reg::Register,
         csrs: Rc<RefCell<csr::CSRs>>,
         bus: bus::Bus,
@@ -42,11 +42,11 @@ impl CPU {
     }
 
     pub fn add2pc(&mut self, addval: i32) {
-        self.pc = (self.pc as i32 + addval) as u32;
+        self.pc = (self.pc as i32 + addval) as usize;
     }
 
     pub fn update_pc(&mut self, newval: i32) {
-        self.pc = newval as u32;
+        self.pc = newval as usize;
     }
 }
